king queen chessboard
given coordinates of positions of king and queen in 8X8 chessboard, program should determine if the king is in check position based on the current positions of the pieces. and if so, return the number of spaces it can move to in order to get out of check.if the king is not in check position return -1.Remember , because only the queen and king are on board, if the queen is checking the king by directly adjacent to it, technically the king can capture the queen.